requirements.txt
psycopg2-binary
kafka-python==2.0.2